requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.3.1
astroid==2.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-celery==3.3.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-seed==0.2.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jsonapi==4.1.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
Faker==8.1.1
filelock==3.0.12
identify==2.2.2
idna==2.10
inflection==0.5.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.3
kombu==3.0.37
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
nodeenv==1.5.0
oauthlib==3.1.0
pre-commit==2.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
virtualenv==20.4.3
wcwidth==0.2.5
wrapt==1.11.2
